+++ models/index.js
@@ -40,6 +40,7 @@ const modelFiles = [
   require('./rxworkflowtracking.js'),
   require('./systemsetting.js'),
   require('./user.js'),
+  require('./useractivitylog.js'),
   require('./workflowaction.js'),
 ];